QPI_Analytics
The QPI Analysis Tool is Property of The QPI Development Team. Quantum Phonetic Integration is 35 (A)111 USC Patent Pending. It is free to use for personal and development purposes only. NOT FOR SALE or commercial usage.  

The QPI Analysis tool accesses TextBlob. and spaCy for access to artificial intelligence NLP learning mechanisms. The current release is version 1.0b and includes minimal intelligence analysis and accessibility to TextBlob and spaCy  features. Advanced features will be available in future versions. We have integrated Hunnspell spelling checks as well as a localized library for adding words and checking them against Hunnspell’s available word catalog. We have added a Bad Word Dictionary to add words that Hunnspell has deemed correctly spelled which aren’t. This feature is currently not working. The Advanced features in the Analysis features are listed but not currently functional. In time they will be.